Apply dhewm3 / 94df46637aac33da7aab9f23874f372fb23c0d2d: s/WriteLong/WriteInt/ to match the argument type
git-svn-id: http://svn.thedarkmod.com/svn/darkmod_src/branches/x64@6292 49c82d7f-2e2a-0410-a16f-ae8f201b507f

+++ framework/async/AsyncClient.cpp
@@ -291,7 +291,7 @@ void idAsyncClient::GetServerInfo( const netadr_t adr ) {
 	msg.Init( msgBuf, sizeof( msgBuf ) );
 	msg.WriteShort( CONNECTIONLESS_MESSAGE_ID );
 	msg.WriteString( "getInfo" );
-	msg.WriteLong( serverList.GetChallenge() );	// challenge
+	msg.WriteInt( serverList.GetChallenge() );	// challenge
 
 	clientPort.SendPacket( adr, msg.GetData(), msg.GetSize() );	
 }
@@ -350,7 +350,7 @@ void idAsyncClient::GetLANServers( void ) {
 	msg.Init( msgBuf, sizeof( msgBuf ) );
 	msg.WriteShort( CONNECTIONLESS_MESSAGE_ID );
 	msg.WriteString( "getInfo" );
-	msg.WriteLong( serverList.GetChallenge() );
+	msg.WriteInt( serverList.GetChallenge() );
 
 	broadcastAddress.type = NA_BROADCAST;
 	for ( i = 0; i < MAX_SERVER_PORTS; i++ ) {
@@ -378,7 +378,7 @@ void idAsyncClient::GetNETServers( void ) {
 	msg.Init( msgBuf, sizeof( msgBuf ) );
 	msg.WriteShort( CONNECTIONLESS_MESSAGE_ID );
 	msg.WriteString( "getServers" );
-	msg.WriteLong( ASYNC_PROTOCOL_VERSION );
+	msg.WriteInt( ASYNC_PROTOCOL_VERSION );
 	msg.WriteString( cvarSystem->GetCVarString( "fs_mod" ) );
 	msg.WriteBits( cvarSystem->GetCVarInteger( "gui_filter_password" ), 2 );
 	msg.WriteBits( cvarSystem->GetCVarInteger( "gui_filter_players" ), 2 );
@@ -606,9 +606,9 @@ void idAsyncClient::SendEmptyToServer( bool force, bool mapLoad ) {
 	}
 
 	msg.Init( msgBuf, sizeof( msgBuf ) );
-	msg.WriteLong( serverMessageSequence );
-	msg.WriteLong( mapLoad ? GAME_INIT_ID_MAP_LOAD : gameInitId );
-	msg.WriteLong( snapshotSequence );
+	msg.WriteInt( serverMessageSequence );
+	msg.WriteInt( mapLoad ? GAME_INIT_ID_MAP_LOAD : gameInitId );
+	msg.WriteInt( snapshotSequence );
 	msg.WriteByte( CLIENT_UNRELIABLE_MESSAGE_EMPTY );
 
 	channel.SendMessage( clientPort, clientTime, msg );
@@ -634,11 +634,11 @@ void idAsyncClient::SendPingResponseToServer( int time ) {
 	}
 
 	msg.Init( msgBuf, sizeof( msgBuf ) );
-	msg.WriteLong( serverMessageSequence );
-	msg.WriteLong( gameInitId );
-	msg.WriteLong( snapshotSequence );
+	msg.WriteInt( serverMessageSequence );
+	msg.WriteInt( gameInitId );
+	msg.WriteInt( snapshotSequence );
 	msg.WriteByte( CLIENT_UNRELIABLE_MESSAGE_PINGRESPONSE );
-	msg.WriteLong( time );
+	msg.WriteInt( time );
 
 	channel.SendMessage( clientPort, clientTime, msg );
 	while( channel.UnsentFragmentsLeft() ) {
@@ -669,16 +669,16 @@ void idAsyncClient::SendUsercmdsToServer( void ) {
 
 	// send the user commands to the server
 	msg.Init( msgBuf, sizeof( msgBuf ) );
-	msg.WriteLong( serverMessageSequence );
-	msg.WriteLong( gameInitId );
-	msg.WriteLong( snapshotSequence );
+	msg.WriteInt( serverMessageSequence );
+	msg.WriteInt( gameInitId );
+	msg.WriteInt( snapshotSequence );
 	msg.WriteByte( CLIENT_UNRELIABLE_MESSAGE_USERCMD );
 	msg.WriteShort( clientPrediction );
 
 	numUsercmds = idMath::ClampInt( 0, 10, idAsyncNetwork::clientUsercmdBackup.GetInteger() ) + 1;
 
 	// write the user commands
-	msg.WriteLong( gameFrame );
+	msg.WriteInt( gameFrame );
 	msg.WriteByte( numUsercmds );
 	for ( last = NULL, i = gameFrame - numUsercmds + 1; i <= gameFrame; i++ ) {
 		index = i & ( MAX_USERCMD_BACKUP - 1 );
@@ -1425,19 +1425,19 @@ void idAsyncClient::SetupConnection( void ) {
 		msg.Init( msgBuf, sizeof( msgBuf ) );
 		msg.WriteShort( CONNECTIONLESS_MESSAGE_ID );
 		msg.WriteString( "challenge" );
-		msg.WriteLong( clientId );
+		msg.WriteInt( clientId );
 		clientPort.SendPacket( serverAddress, msg.GetData(), msg.GetSize() );
 	} else if ( clientState == CS_CONNECTING ) {
 		common->Printf( "sending connect to %s with challenge 0x%x\n", Sys_NetAdrToString( serverAddress ), serverChallenge );
 		msg.Init( msgBuf, sizeof( msgBuf ) );
 		msg.WriteShort( CONNECTIONLESS_MESSAGE_ID );
 		msg.WriteString( "connect" );
-		msg.WriteLong( ASYNC_PROTOCOL_VERSION );
+		msg.WriteInt( ASYNC_PROTOCOL_VERSION );
 		msg.WriteShort( BUILD_OS_ID ); // 0 to fake win32
-		msg.WriteLong( clientDataChecksum );
-		msg.WriteLong( serverChallenge );
+		msg.WriteInt( clientDataChecksum );
+		msg.WriteInt( serverChallenge );
 		msg.WriteShort( clientId );
-		msg.WriteLong( cvarSystem->GetCVarInteger( "net_clientMaxRate" ) );
+		msg.WriteInt( cvarSystem->GetCVarInteger( "net_clientMaxRate" ) );
 		msg.WriteString( cvarSystem->GetCVarString( "com_guid" ) );
 		msg.WriteString( cvarSystem->GetCVarString( "password" ), -1, false );
 		// do not make the protocol depend on PB
@@ -1452,7 +1452,7 @@ void idAsyncClient::SetupConnection( void ) {
 			msg.BeginWriting();
 			msg.WriteShort( CONNECTIONLESS_MESSAGE_ID );
 			msg.WriteString( "clAuth" );
-			msg.WriteLong( ASYNC_PROTOCOL_VERSION );
+			msg.WriteInt( ASYNC_PROTOCOL_VERSION );
 			msg.WriteNetadr( serverAddress );
 			// if we don't have a com_guid, this will request a direct reply from auth with it
 			msg.WriteByte( cvarSystem->GetCVarString( "com_guid" )[0] ? 1 : 0 );
@@ -1671,7 +1671,7 @@ void idAsyncClient::SendVersionCheck( bool fromMenu ) {
 	msg.Init( msgBuf, sizeof( msgBuf ) );
 	msg.WriteShort( CONNECTIONLESS_MESSAGE_ID );
 	msg.WriteString( "versionCheck" );
-	msg.WriteLong( ASYNC_PROTOCOL_VERSION );
+	msg.WriteInt( ASYNC_PROTOCOL_VERSION );
 	msg.WriteShort( BUILD_OS_ID );
 	msg.WriteString( cvarSystem->GetCVarString( "si_version" ) );
 	msg.WriteString( cvarSystem->GetCVarString( "com_guid" ) );
@@ -1700,7 +1700,7 @@ void idAsyncClient::SendVersionDLUpdate( int state ) {
 	msg.Init( msgBuf, sizeof( msgBuf ) );
 	msg.WriteShort( CONNECTIONLESS_MESSAGE_ID );
 	msg.WriteString( "versionDL" );
-	msg.WriteLong( ASYNC_PROTOCOL_VERSION );
+	msg.WriteInt( ASYNC_PROTOCOL_VERSION );
 	msg.WriteShort( state );
 	clientPort.SendPacket( idAsyncNetwork::GetMasterAddress(), msg.GetData(), msg.GetSize() );
 }
@@ -1910,7 +1910,7 @@ idAsyncClient::SendAuthCheck
 	msg.Init( msgBuf, sizeof( msgBuf ) );
 	msg.WriteShort( CONNECTIONLESS_MESSAGE_ID );
 	msg.WriteString( "gameAuth" );
-	msg.WriteLong( ASYNC_PROTOCOL_VERSION );
+	msg.WriteInt( ASYNC_PROTOCOL_VERSION );
 	msg.WriteByte( cdkey ? 1 : 0 );
 	msg.WriteString( cdkey ? cdkey : "" );
 	msg.WriteByte( xpkey ? 1 : 0 );
