+++ game/Player.cpp
@@ -9268,7 +9268,7 @@ idPlayer::WritePlayerStateToSnapshot
 */
 void idPlayer::WritePlayerStateToSnapshot( idBitMsgDelta &msg ) const {
 	msg.WriteByte( bobCycle );
-	msg.WriteLong( stepUpTime );
+	msg.WriteInt( stepUpTime );
 	msg.WriteFloat( stepUpDelta );
 }
